Initial commit - MSVC
Initial commit of files that have been compiled and run successfully in MSVC2015. 

You're welcome to try compiling with GCC5, as it requires std::atomic_exchange implemented for shared_ptrs.  Using -std=c++14 is desirable.  Full support and makefiles directly working with GCC will be the next course of updates to this library.